requirements.txt
boto3
joblib
keras==2.3.1
tensorflow==2.1.0
nltk
numpy
pymongo
requests
python-dotenv